// src/lib/types.ts
export interface UsersType {
    user_id: number;
    user_name: string;
    user_fullname: string;
    user_state: string;
}

export interface EmailsType {
    id: number,
    brand_name: string;
    brand_domain: string;
    state: string;
    note: string;
    employees: string
    updated_at: string;
    created_at: string;
}